ai_chatbot/requirements.txt
Flask>=3.0,<4.0
flask-cors>=4.0,<5.0
mysql-connector-python>=8.3,<10.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
scikit-learn>=1.3.0
matplotlib>=3.7.0,<4.0